+++ src/explore-education-statistics-admin/src/pages/publication/PublicationReleasesPage.module.scss
@@ -6,6 +6,7 @@
   }
 }
 
+.actionsColumn,
 .statusColumn {
   width: 150px;
 }
