+++ src/explore-education-statistics-admin/src/pages/release/components/ReleasePublishingStages.tsx
@@ -1,0 +1,71 @@
+import StatusBlock from '@admin/components/StatusBlock';
+import getStatusDetail from '@admin/pages/release/utils/releaseStatusDetail';
+import { ReleaseStageStatuses } from '@admin/services/releaseService';
+import Details from '@common/components/Details';
+import React from 'react';
+
+const notStartedStatuses = ['Validating', 'Invalid'];
+
+interface Props {
+  checklistStyle?: boolean;
+  currentStatus?: ReleaseStageStatuses;
+  includeScheduled?: boolean;
+}
+
+const ReleasePublishingStages = ({
+  checklistStyle = false,
+  currentStatus,
+  includeScheduled = false,
+}: Props) => {
+  if (
+    !currentStatus ||
+    (includeScheduled
+      ? notStartedStatuses.includes(currentStatus.overallStage)
+      : [...notStartedStatuses, 'Scheduled'].includes(
+          currentStatus.overallStage,
+        ))
+  ) {
+    return null;
+  }
+  return (
+    <Details
+      className="govuk-!-margin-bottom-0 govuk-!-margin-top-1"
+      summary="View stages"
+    >
+      {checklistStyle &&
+        ![...notStartedStatuses, 'Scheduled'].includes(
+          currentStatus.overallStage,
+        ) && (
+          <p className="govuk-!-font-weight-bold">Release process started</p>
+        )}
+      <ul className="govuk-list">
+        {Object.entries(currentStatus).map(([key, val]) => {
+          if (['overallStage', 'releaseId', 'lastUpdated'].includes(key)) {
+            return null;
+          }
+          const { color, text } = getStatusDetail(val);
+
+          if (!color) {
+            return null;
+          }
+
+          return (
+            <li key={key}>
+              <StatusBlock
+                checklistStyle={checklistStyle}
+                color={color}
+                text={
+                  checklistStyle
+                    ? `${key.replace('Stage', '')} ${text}`
+                    : `${key.replace('Stage', '')} - ${text}`
+                }
+              />
+            </li>
+          );
+        })}
+      </ul>
+    </Details>
+  );
+};
+
+export default ReleasePublishingStages;
